// ui/src/data/generated/TierList.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DataSetInfo } from "./DataSetInfo";
import type { TierEntry } from "./TierEntry";

/**
 * `{patch}/{queue}/{bracket}/tierlist.json`.
 */
export type TierList = { info: DataSetInfo, 
/**
 * Best score first.
 */
entries: Array<TierEntry>, };